refactor: `dict` モジュールへ移植
