+++ test/user_dict/test_user_dict.py
@@ -7,9 +7,8 @@ from unittest import TestCase
 from fastapi import HTTPException
 from pyopenjtalk import g2p, unset_user_dict
 
-from voicevox_engine.model import UserDictWord, WordTypes
-from voicevox_engine.part_of_speech_data import MAX_PRIORITY, part_of_speech_data
-from voicevox_engine.user_dict import (
+from voicevox_engine.dict.part_of_speech_data import MAX_PRIORITY, part_of_speech_data
+from voicevox_engine.dict.user_dict import (
     _create_word,
     apply_word,
     delete_word,
@@ -18,6 +17,7 @@ from voicevox_engine.user_dict import (
     rewrite_word,
     update_dict,
 )
+from voicevox_engine.model import UserDictWord, WordTypes
 
 # jsonとして保存される正しい形式の辞書データ
 valid_dict_dict_json = {
